refactor(web): FilterPanel をディレクトリ化し責務ごとにコンポーネント分割
filter-panel/ 配下に状態・ロジックのコンテナ（index）、mobile/desktop の表示、
共通の FilterControls、FilterField、Filters 型を分離。表示は純粋・ロジックはコンテナに集約し、
mobile/desktop で重複していたフィルタ行を FilterControls に共通化した。
