+++ application/packages/web/src/client/features/article/components/filter-panel/index.tsx
@@ -1,0 +1,89 @@
+import { useState } from 'react'
+import { useIsMobile } from '@/client/components/shadcn/hooks/use-mobile'
+import { scrollToTop } from '@/client/lib/scroll'
+import { type DatePresetType } from '../../hooks/use-articles'
+import { type MediaType } from '../media-filter'
+import { type ReadStatusType } from '../read-status-filter'
+import { DesktopFilterPanel } from './desktop-filter-panel'
+import { MobileFilterPanel } from './mobile-filter-panel'
+import { type Filters } from './types'
+
+const DEFAULT_FILTERS: Filters = { media: null, readStatus: 'all', datePreset: 'today' }
+
+interface FilterPanelProps {
+  selectedMedia: MediaType
+  selectedReadStatus: ReadStatusType
+  selectedDatePreset: DatePresetType
+  onApplyFilters: (filters: Filters) => void
+  isLoggedIn: boolean
+}
+
+export function FilterPanel({
+  selectedMedia,
+  selectedReadStatus,
+  selectedDatePreset,
+  onApplyFilters,
+  isLoggedIn,
+}: FilterPanelProps) {
+  const isMobile = useIsMobile()
+  const applied: Filters = {
+    media: selectedMedia,
+    readStatus: selectedReadStatus,
+    datePreset: selectedDatePreset,
+  }
+  const [isFilterOpen, setIsFilterOpen] = useState(false)
+  const [draft, setDraft] = useState<Filters>(applied)
+
+  const appliedFilterCount =
+    (applied.media ? 1 : 0) +
+    (isLoggedIn && applied.readStatus === 'unread' ? 1 : 0) +
+    (applied.datePreset !== 'today' ? 1 : 0)
+
+  const commitFilters = (filters: Filters) => {
+    onApplyFilters(filters)
+    setIsFilterOpen(false)
+    scrollToTop()
+  }
+
+  const editDraft = (patch: Partial<Filters>) => {
+    setDraft((current) => ({ ...current, ...patch }))
+  }
+
+  const commitDraft = (patch: Partial<Filters>) => {
+    const next = { ...draft, ...patch }
+    setDraft(next)
+    commitFilters(next)
+  }
+
+  const toggleFilter = () => {
+    if (isFilterOpen) {
+      setIsFilterOpen(false)
+      return
+    }
+    setDraft(applied)
+    setIsFilterOpen(true)
+  }
+
+  const cancelFilter = () => {
+    setDraft(applied)
+    setIsFilterOpen(false)
+  }
+
+  if (isMobile) {
+    return (
+      <MobileFilterPanel
+        isOpen={isFilterOpen}
+        appliedFilterCount={appliedFilterCount}
+        draft={draft}
+        isLoggedIn={isLoggedIn}
+        onToggle={toggleFilter}
+        onDraftChange={editDraft}
+        onCancel={cancelFilter}
+        onClear={() => commitDraft(DEFAULT_FILTERS)}
+        onApply={() => commitFilters(draft)}
+      />
+    )
+  }
+
+  return <DesktopFilterPanel draft={draft} isLoggedIn={isLoggedIn} onCommit={commitDraft} />
+}
